+++ Controller/ListRegularizacionImpuesto.php
@@ -46,61 +46,50 @@ class ListRegularizacionImpuesto extends ListController
     protected function createViews(): void
     {
         $this->createViewsModel303();
-        $this->createViewsModel303New();
         $this->createViewsModel390();
     }
 
+    /**
+     * Create the list view for Model 303.
+     *
+     * @param string $viewName
+     * @return void
+     */
     protected function createViewsModel303(string $viewName = 'ListRegularizacionImpuesto'): void
     {
+        $exercises = $this->codeModel->all('ejercicios', 'codejercicio', 'nombre');
         $this->addView($viewName, 'RegularizacionImpuesto', 'model-303', 'fa-solid fa-book')
+            // Search and Orderby
+            ->addSearchFields(['codsubcuentaacr', 'codsubcuentadeu'])
             ->addOrderBy(['fechainicio'], 'start-date', 2)
             ->addOrderBy(['codejercicio||periodo'], 'period')
-            ->addSearchFields(['codsubcuentaacr', 'codsubcuentadeu']);
-
-        // añadimos filtros
-        $this->addFilterSelectWhere($viewName, 'status', [
-            ['label' => Tools::lang()->trans('model-303'), 'where' => [new DataBaseWhere('periodo', 'Y', '!=')]]
-        ]);
-
-        $this->addFilterSelect($viewName, 'idempresa', 'company', 'idempresa', Empresas::codeModel());
-
-        $exercises = $this->codeModel->all('ejercicios', 'codejercicio', 'nombre');
-        $this->addFilterSelect($viewName, 'codejercicio', 'exercise', 'codejercicio', $exercises);
+            // Filters
+            ->addFilterSelect('idempresa', 'company', 'idempresa', Empresas::codeModel())
+            ->addFilterSelect('codejercicio', 'exercise', 'codejercicio', $exercises)
+            ->addFilterSelectWhere('status', [
+                ['label' => Tools::lang()->trans('model-303'), 'where' => [new DataBaseWhere('periodo', 'Y', '!=')]]
+            ]);
     }
 
-    protected function createViewsModel303New(string $viewName = 'ListRegularizacionImpuesto-New'): void
-    {
-        $this->addView($viewName, 'RegularizacionImpuestoNew', 'model-303-new', 'fa-solid fa-book')
-            ->addOrderBy(['fechainicio'], 'start-date', 2)
-            ->addOrderBy(['codejercicio||periodo'], 'period')
-            ->addSearchFields(['codsubcuentaacr', 'codsubcuentadeu']);
-
-        // añadimos filtros
-        $this->addFilterSelectWhere($viewName, 'status', [
-            ['label' => Tools::lang()->trans('model-303'), 'where' => [new DataBaseWhere('periodo', 'Y', '!=')]]
-        ]);
-
-        $this->addFilterSelect($viewName, 'idempresa', 'company', 'idempresa', Empresas::codeModel());
-
-        $exercises = $this->codeModel->all('ejercicios', 'codejercicio', 'nombre');
-        $this->addFilterSelect($viewName, 'codejercicio', 'exercise', 'codejercicio', $exercises);
-    }
-
+    /**
+     * Create the list view for Model 390.
+     *
+     * @param string $viewName
+     * @return void
+     */
     protected function createViewsModel390(string $viewName = 'ListRegularizacionImpuesto-390'): void
     {
+        $exercises = $this->codeModel->all('ejercicios', 'codejercicio', 'nombre');
         $this->addView($viewName, 'RegularizacionImpuesto', 'model-390', 'fa-solid fa-book')
+            // Search and Orderby
             ->addOrderBy(['fechainicio'], 'start-date', 2)
             ->addOrderBy(['codejercicio||periodo'], 'period')
-            ->addSearchFields(['codsubcuentaacr', 'codsubcuentadeu']);
-
-        // añadimos filtros
-        $this->addFilterSelectWhere($viewName, 'status', [
-            ['label' => Tools::lang()->trans('model-390'), 'where' => [new DataBaseWhere('periodo', 'Y')]]
-        ]);
-
-        $this->addFilterSelect($viewName, 'idempresa', 'company', 'idempresa', Empresas::codeModel());
-
-        $exercises = $this->codeModel->all('ejercicios', 'codejercicio', 'nombre');
-        $this->addFilterSelect($viewName, 'codejercicio', 'exercise', 'codejercicio', $exercises);
+            ->addSearchFields(['codsubcuentaacr', 'codsubcuentadeu'])
+            // Filters
+            ->addFilterSelect('idempresa', 'company', 'idempresa', Empresas::codeModel())
+            ->addFilterSelect('codejercicio', 'exercise', 'codejercicio', $exercises)
+            ->addFilterSelectWhere('status', [
+                ['label' => Tools::lang()->trans('model-390'), 'where' => [new DataBaseWhere('periodo', 'Y')]]
+            ]);
     }
 }
