--- XMLView/ListPartidaImpuestoResumenNew.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
- * This file is part of Modelo303 plugin for FacturaScripts
- * Copyright (C) 2017-2022 Carlos Garcia Gomez <carlos@facturascripts.com>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- *
- * @author Jose Antonio Cuello Principal <yopli2000@gmail.com>
--->
-<view>
-    <columns>
-        <column name="special-account" order="100">
-            <widget type="text" fieldname="codcuentaesp"/>
-        </column>
-        <column name="subaccount" order="110">
-            <widget type="text" fieldname="codsubcuenta" />
-        </column>
-        <column name="description" order="120">
-            <widget type="textarea" fieldname="descripcion"/>
-        </column>
-        <column name="pct-vat" display="right" order="130">
-            <widget type="number" decimal="2" fieldname="iva"/>
-        </column>
-        <column name="pct-surcharge" display="right" order="140">
-            <widget type="number" decimal="2" fieldname="recargo"/>
-        </column>
-        <column name="tax-base" display="right" order="150">
-            <widget type="money" fieldname="baseimponible">
-                <option color="danger">lt:0</option>
-            </widget>
-        </column>
-        <column name="quota" display="right" order="160">
-            <widget type="money" fieldname="cuota">
-                <option color="danger">lt:0</option>
-            </widget>
-        </column>
-    </columns>
-</view>
